fix(react-portal): stop click event propagation
Components that uses react-portal open/close state and dismisses when
clicked outside of the menu that pops up. In react 16, the event is
captured after mount, but in react 17 the event is captured due to
changes in the react event listener architecture. This causes the menu
to close immediately after being open resulting in nothing happening.

Instead of relying on a side effect to occur to dismiss the popup,
change it to explicitly toggle the menu and stop the event from
propagating up.