fix(small-footer): explicitly close the menu when in the copyright state
Previously the small footer button is clicked, it relies on an event
listener to close. For the expanded state, it was explicitly closed. Due
to react 17's event listener propagation changes, we no longer receive
the event in the same order as in react 16. Instead, explicitly close
the small footer menu.